refactor: move mergeSubStyle function to styleUtils and update its usage in markdown style files
